fix: Remove comment syntax match
As the todo: `Markdown and code doc comments` is not complete. Comment
section was removed as it was overriding the default rust syntax
group that currently works. In the future this can be re-added when the
feature is complete. As it is now it should not override functionality
that currently works.

Fix #5